test(validation): split the shipped-route suite by concern
The suite reached 568 lines, past the 500-line boundary taste-lints enforces,
because the AST rewrite added seventeen cases. Splitting it by concern rather
than by size: one file pins the invariant the gate exists to enforce (routes
resolve, drift is reported with file and line, unscannable input is a config
error), the other pins which text counts as a route at all (fences, comments,
indented and inline code, blockquotes, list-indented and pipeless tables).

Shared builders move to a helpers module rather than the package conftest,
because a fixture named `repo` in the package-wide conftest would be visible to
every validation suite.

Mutation coverage is unchanged: the same eleven mutations run against the split
suite, ten are caught, and the survivor is the performance prefilter that by
construction changes no outcome.

Refs #3716
